Add an xml string, text paragraph, or gt object at a specified position in a Word document

Wrappers for officer::body_add_par() and officer::body_add_xml() that use the cursor_docx() helper functions to allow input of filename and path and use of multiple options for cursor placement. add_text_to_body() passes value to glue::glue() to add support for glue string interpolation. add_gt_to_body() converts gt tables to OOXML with gt::as_word(). If pos = NULL, add_to_body() calls officer::body_add() instead of officer::body_add_par().

Usage

add_to_body(
  docx,
  keyword = NULL,
  id = NULL,
  index = NULL,
  value = NULL,
  str = NULL,
  pos = "after",
  ...
)

add_text_to_body(
  docx,
  value,
  style = NULL,
  pos = "after",
  .na = "NA",
  .null = NULL,
  .envir = parent.frame(),
  ...
)

add_xml_to_body(docx, str, pos = "after", ...)

add_gt_to_body(docx, gt_object, pos = "after", ...)

add_value_with_keys(docx, value, ..., .f = add_text_to_body)

add_str_with_keys(docx, str, ..., .f = add_xml_to_body)

Arguments

docx

A rdocx object.

keyword, id

A keyword string used to place cursor with officer::cursor_reach() or bookmark id with officer::cursor_bookmark(). Defaults to NULL. If keyword or id are not provided, the gt object is inserted at the front of the document.

index

A integer matching a doc_index value appearing in a summary of the docx object created with officer::docx_summary(). If index is for a paragraph value, the text of the pargraph is used as a keyword.

value

object to add in the document. Supported objects are vectors, data.frame, graphics, block of formatted paragraphs, unordered list of formatted paragraphs, pretty tables with package flextable, 'Microsoft' charts with package mschart.

str

a wml string

pos

where to add the new element relative to the cursor, one of "after", "before", "on".

...

Additional parameters passed to officer::body_add() or officer::body_add_xml().

style

paragraph style name. These names are available with function styles_info and are the names of the Word styles defined in the base document (see argument path from read_docx).

.na

[character(1): ‘NA’]
Value to replace NA values with. If NULL missing values are propagated, that is an NA result will cause NA output. Otherwise the value is replaced by the value of .na.

.null

[character(1): ‘character()’]
Value to replace NULL values with. If character() whole output is character(). If NULL all NULL values are dropped (as in paste0()). Otherwise the value is replaced by the value of .null.

.envir

[environment: parent.frame()]
Environment to evaluate each expression in. Expressions are evaluated from left to right. If .x is an environment, the expressions are evaluated in that environment and .envir is ignored. If NULL is passed, it is equivalent to emptyenv().

gt_object

A gt object converted to an OOXML string with gt::as_word() then passed to add_xml_to_body() as str parameter. Required for add_gt_to_body().

.f

Any function that takes a docx and value parameter and returns a rdocx object. A keyword parameter must also be supported if named is TRUE. Defaults to add_text_to_body().

Value

A rdocx object with added xml, gt tables, or paragraphs of text.

Details

Using add_value_with_keys() or add_str_with_keys()

add_value_with_keys() supports value vectors of length 1 or longer. If value is named, the names are assumed to be keywords indicating the cursor position for adding each value in the vector. If value is not named, a keyword parameter with the same length as value must be provided. When named = FALSE, no keyword parameter is required. Add add_str_with_keys() works identically but uses a str parameter and .f defaults to add_xml_to_body().