Get tables from a rdocx or rpptx object

Get one or more tables from a rdocx or rpptx object. Functions based on example code on extracting Word document and PowerPoint slides in the officeverse documentation.

Usage

officer_tables(
  x,
  index = NULL,
  has_header = TRUE,
  col = NULL,
  ...,
  stack = FALSE,
  call = caller_env()
)

officer_table(
  x,
  index = NULL,
  has_header = TRUE,
  col = NULL,
  ...,
  call = caller_env()
)

Arguments

x

A rdocx or rpptx object or a data.frame created with officer_summary()

index

A index value matching a doc_index value for a table in the summary data.frame, Default: NULL

has_header

If TRUE, tables are expected to have implicit headers even if the object Default: TRUE

col

If col is supplied, officer_table() passes col and the additional parameters in ... to fill_with_pattern(). This allows the addition of preceding headings or captions as a column within the data.frame returned by officer_tables(). This is an experimental feature and may be modified or removed. Defaults to NULL.

...

Additional parameters passed to fill_with_pattern().

stack

If TRUE and all tables share the same number of columns, return a combined data.frame instead of a list object.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

Value

A data.frame or list of data.frame (or list objects).

Details

officer_table() is a lower-level helper fucntion to extract a single table from a document.

Examples

docx <- read_docx_ext(
  filename = "example.docx",
  path = system.file("doc_examples", package = "officer")
)

officer_tables(docx)
#>         Petals   Internode                 Sepal
#> 2  5,621498349        <NA> 2,46210657918,2034091
#> 3  4,994616997          AA           2,429320759
#> 4  4,767504884        <NA>                   AAA
#> 5   25,9242382        <NA>           2,066051345
#> 6  6,489375001 25,21130805           2,901582763
#> 7    5,7858682 25,52433147           2,655642742
#> 8  5,645575295 Merged cell           2,278691288
#> 9  4,828953215        <NA>           2,238467716
#> 10 6,783500773        <NA>           2,202762147
#> 11 5,395076839        <NA>           2,538375992
#> 12 4,683617783  29,2459239           2,601945544
#> 13        Note        <NA>                  <NA>
#>                               Bract
#> 2                              <NA>
#> 3                       17,65204912
#> 4                              <NA>
#> 5                       18,37915478
#> 6  17,3130473717,0721572418,2902189
#> 7                              <NA>
#> 8                              <NA>
#> 9                       19,87376227
#> 10                      19,85326662
#> 11                      19,56545356
#> 12                      18,95335451
#> 13                             <NA>

pptx <- read_pptx_ext(
  filename = "example.pptx",
  path = system.file("doc_examples", package = "officer")
)

officer_tables(pptx)[[1]]
#>   Header 1  Header 2       Header 3
#> 2         A    12.23      blah blah
#> 3         B     1.23 blah blah blah
#> 4         B      9.0          Salut
#> 5         C        6          Hello