Add an xml string, text paragraph, or gt object at a specified position in a rdocx object

Wrappers for officer::body_add_par(), officer::body_add_gg(), and officer::body_add_xml() that use the cursor_docx() helper function to allow users to pass the value and keyword, id, or index value used to place a "cursor" within the document using a single function. If pos = NULL, add_to_body() calls officer::body_add() instead of officer::body_add_par().

add_text_to_body() passes value to glue::glue() to add support for glue string interpolation.

add_gt_to_body() converts gt tables to OOXML with gt::as_word().

add_gg_to_body() adds a caption following the plots using the labels from the plot object.

Usage

add_to_body(
  docx,
  keyword = NULL,
  id = NULL,
  index = NULL,
  value = NULL,
  str = NULL,
  pos = "after",
  call = parent.frame(),
  ...
)

add_text_to_body(
  docx,
  value,
  style = NULL,
  pos = "after",
  .na = "NA",
  .null = NULL,
  .envir = parent.frame(),
  ...
)

add_xml_to_body(docx, str, pos = "after", ...)

add_gt_to_body(
  docx,
  gt_object,
  align = "center",
  caption_location = c("top", "bottom", "embed"),
  caption_align = "left",
  split = FALSE,
  keep_with_next = TRUE,
  pos = "after",
  ...
)

add_gg_to_body(
  docx,
  value,
  caption = "title",
  caption_style = style,
  autonum = NULL,
  style = "Normal",
  pos = "after",
  ...
)

add_value_with_keys(docx, value, ..., .f = add_text_to_body)

add_str_with_keys(docx, str, ..., .f = add_xml_to_body)

Arguments

docx

A rdocx object.

keyword, id

A keyword string used to place cursor with officer::cursor_reach() or bookmark id with officer::cursor_bookmark(). Defaults to NULL. If keyword or id are not provided, the gt object is inserted at the front of the document.

index

A integer matching a doc_index value appearing in a summary of the docx object created with officer::docx_summary(). If index is for a paragraph value, the text of the pargraph is used as a keyword.

value

object to add in the document. Supported objects are vectors, data.frame, graphics, block of formatted paragraphs, unordered list of formatted paragraphs, pretty tables with package flextable, 'Microsoft' charts with package mschart.

str

a wml string

pos

where to add the new element relative to the cursor, one of "after", "before", "on".

call

The execution environment of a currently running function passed to cli::cli_abort(). Defaults to parent.frame()

...

Additional parameters passed to officer::body_add_par(), officer::body_add_gg(), or officer::body_add().

style

paragraph style name. These names are available with function styles_info and are the names of the Word styles defined in the base document (see argument path from read_docx).

.na

[character(1): ‘NA’]
Value to replace NA values with. If NULL missing values are propagated, that is an NA result will cause NA output. Otherwise the value is replaced by the value of .na.

.null

[character(1): ‘character()’]
Value to replace NULL values with. If character() whole output is character(). If NULL all NULL values are dropped (as in paste0()). Otherwise the value is replaced by the value of .null.

.envir

[environment: parent.frame()]
Environment to evaluate each expression in. Expressions are evaluated from left to right. If .x is an environment, the expressions are evaluated in that environment and .envir is ignored. If NULL is passed, it is equivalent to emptyenv().

gt_object

A gt object converted to an OOXML string with gt::as_word() then passed to add_xml_to_body() as str parameter. Required for add_gt_to_body().

align

An option for table alignment. Can either be "center" (the default), "left", or "right".

caption_location

Determines where the caption should be positioned. This can either be "top" (the default), "bottom", or "embed".

caption_align

Determines the alignment of the caption. This is either "left" (the default), "center", or "right". This option is only used when caption_location is not set as "embed".

split

A TRUE or FALSE (the default) value that indicates whether to activate the Word option Allow row to break across pages.

keep_with_next

A TRUE (the default) or FALSE value that indicates whether a table should use Word option keep rows together.

caption

Name of the ggplot2 label to use as a caption if plot passed to value has a label for this value. Defaults to "title".

caption_style

Passed to style for officer::body_add_caption(). Defaults to same value as style.

.f

Any function that takes a docx and value parameter and returns a rdocx object. A keyword parameter must also be supported if named is TRUE. Defaults to add_text_to_body().

Value

A rdocx object with xml, gt tables, or paragraphs of text added.

Details

Using add_value_with_keys() or add_str_with_keys()

add_value_with_keys() supports value vectors of length 1 or longer. If value is named, the names are assumed to be keywords indicating the cursor position for adding each value in the vector. If value is not named, a keyword parameter with the same length as value must be provided. When named = FALSE, no keyword parameter is required. Add add_str_with_keys() works identically but uses a str parameter and .f defaults to add_xml_to_body().

Examples

if (rlang::is_installed("gt")) {
  docx <- read_docx_ext(
    filename = "example.docx",
    path = system.file("doc_examples", package = "officer")
  )

  tab_1 <-
    gt::gt(
      gt::exibble,
      rowname_col = "row",
      groupname_col = "group"
    )

  add_gt_to_body(
    docx,
    tab_1,
    keyword = "Title 1"
  )

  tab_str <- gt::as_word(tab_1)

  add_str_with_keys(
    docx,
    str = c("Title 1" = tab_str, "Title 2" = tab_str)
  )
}
#> rdocx document with 23 element(s)
#> 
#> * styles:
#>                 Normal              heading 1              heading 2 
#>            "paragraph"            "paragraph"            "paragraph" 
#> Default Paragraph Font           Normal Table                No List 
#>            "character"                "table"            "numbering" 
#>            Titre 1 Car            Titre 2 Car          Light Shading 
#>            "character"            "character"                "table" 
#>         List Paragraph                 header            En-tête Car 
#>            "paragraph"            "paragraph"            "character" 
#>                 footer       Pied de page Car 
#>            "paragraph"            "character" 
#> 
#> * Content at cursor location:
#>      row_id is_header cell_id             text col_span row_span content_type
#> 1.1       1      TRUE       1                         1        1   table cell
#> 1.9       2     FALSE       1            grp_a        1        1   table cell
#> 1.10      3     FALSE       1            row_1        1        1   table cell
#> 1.18      4     FALSE       1            row_2        1        1   table cell
#> 1.26      5     FALSE       1            row_3        1        1   table cell
#> 1.34      6     FALSE       1            row_4        1        1   table cell
#> 1.42      7     FALSE       1            grp_b        1        1   table cell
#> 1.43      8     FALSE       1            row_5        1        1   table cell
#> 1.51      9     FALSE       1            row_6        1        1   table cell
#> 1.59     10     FALSE       1            row_7        1        1   table cell
#> 1.67     11     FALSE       1            row_8        1        1   table cell
#> 2.2       1      TRUE       2              num        1        1   table cell
#> 2.11      3     FALSE       2        1.111e-01        1        1   table cell
#> 2.19      4     FALSE       2        2.222e+00        1        1   table cell
#> 2.27      5     FALSE       2        3.333e+01        1        1   table cell
#> 2.35      6     FALSE       2        4.444e+02        1        1   table cell
#> 2.44      8     FALSE       2        5.550e+03        1        1   table cell
#> 2.52      9     FALSE       2               NA        1        1   table cell
#> 2.60     10     FALSE       2        7.770e+05        1        1   table cell
#> 2.68     11     FALSE       2        8.880e+06        1        1   table cell
#> 3.3       1      TRUE       3             char        1        1   table cell
#> 3.12      3     FALSE       3          apricot        1        1   table cell
#> 3.20      4     FALSE       3           banana        1        1   table cell
#> 3.28      5     FALSE       3          coconut        1        1   table cell
#> 3.36      6     FALSE       3           durian        1        1   table cell
#> 3.45      8     FALSE       3               NA        1        1   table cell
#> 3.53      9     FALSE       3              fig        1        1   table cell
#> 3.61     10     FALSE       3       grapefruit        1        1   table cell
#> 3.69     11     FALSE       3         honeydew        1        1   table cell
#> 4.4       1      TRUE       4             fctr        1        1   table cell
#> 4.13      3     FALSE       4              one        1        1   table cell
#> 4.21      4     FALSE       4              two        1        1   table cell
#> 4.29      5     FALSE       4            three        1        1   table cell
#> 4.37      6     FALSE       4             four        1        1   table cell
#> 4.46      8     FALSE       4             five        1        1   table cell
#> 4.54      9     FALSE       4              six        1        1   table cell
#> 4.62     10     FALSE       4            seven        1        1   table cell
#> 4.70     11     FALSE       4            eight        1        1   table cell
#> 5.5       1      TRUE       5             date        1        1   table cell
#> 5.14      3     FALSE       5       2015-01-15        1        1   table cell
#> 5.22      4     FALSE       5       2015-02-15        1        1   table cell
#> 5.30      5     FALSE       5       2015-03-15        1        1   table cell
#> 5.38      6     FALSE       5       2015-04-15        1        1   table cell
#> 5.47      8     FALSE       5       2015-05-15        1        1   table cell
#> 5.55      9     FALSE       5       2015-06-15        1        1   table cell
#> 5.63     10     FALSE       5               NA        1        1   table cell
#> 5.71     11     FALSE       5       2015-08-15        1        1   table cell
#> 6.6       1      TRUE       6             time        1        1   table cell
#> 6.15      3     FALSE       6            13:35        1        1   table cell
#> 6.23      4     FALSE       6            14:40        1        1   table cell
#> 6.31      5     FALSE       6            15:45        1        1   table cell
#> 6.39      6     FALSE       6            16:50        1        1   table cell
#> 6.48      8     FALSE       6            17:55        1        1   table cell
#> 6.56      9     FALSE       6               NA        1        1   table cell
#> 6.64     10     FALSE       6            19:10        1        1   table cell
#> 6.72     11     FALSE       6            20:20        1        1   table cell
#> 7.7       1      TRUE       7         datetime        1        1   table cell
#> 7.16      3     FALSE       7 2018-01-01 02:22        1        1   table cell
#> 7.24      4     FALSE       7 2018-02-02 14:33        1        1   table cell
#> 7.32      5     FALSE       7 2018-03-03 03:44        1        1   table cell
#> 7.40      6     FALSE       7 2018-04-04 15:55        1        1   table cell
#> 7.49      8     FALSE       7 2018-05-05 04:00        1        1   table cell
#> 7.57      9     FALSE       7 2018-06-06 16:11        1        1   table cell
#> 7.65     10     FALSE       7 2018-07-07 05:22        1        1   table cell
#> 7.73     11     FALSE       7               NA        1        1   table cell
#> 8.8       1      TRUE       8         currency        1        1   table cell
#> 8.17      3     FALSE       8           49.950        1        1   table cell
#> 8.25      4     FALSE       8           17.950        1        1   table cell
#> 8.33      5     FALSE       8            1.390        1        1   table cell
#> 8.41      6     FALSE       8        65100.000        1        1   table cell
#> 8.50      8     FALSE       8         1325.810        1        1   table cell
#> 8.58      9     FALSE       8           13.255        1        1   table cell
#> 8.66     10     FALSE       8               NA        1        1   table cell
#> 8.74     11     FALSE       8            0.440        1        1   table cell