Make a list of blocks to add to a Word document or PowerPoint presentation

make_block_list() extends officer::block_list() by supporting a list of inputs and optionally combining parameters with an existing block list (using the blocks parameter). Unlike officer::block_list(), make_block_list() errors if no input parameters are provided.

combine_blocks() takes any number of block_list objects and combined them into a single block_list. Both functions are not yet working as expected.

Usage

make_block_list(blocks = NULL, ..., allow_empty = FALSE, call = caller_env())

combine_blocks(...)

Arguments

blocks

A list of parameters to pass to officer::block_list() or a block_list object. If parameters are provided to both ... and blocks is a block_list, the additional parameters are appended to the end of blocks.

...

For make_block_list(), these parameters are passed to officer::block_list() and must not include block_list objects. For combine_blocks(), these parameters must all be block_list objects.

allow_empty

If TRUE, check_block_list() allows an empty block list. Defaults to FALSE.

call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

See also